mics

+++ channels/bank/types.go
@@ -4,9 +4,9 @@ import uuid "github.com/nu7hatch/gouuid"
 
 type Bank interface {
 	CreateAccount() uuid.UUID
-	Deposit(account uuid.UUID, amount float64) float64
-	Withdraw(account uuid.UUID, amount float64) float64
-	Balance(account uuid.UUID) float64
-	Transfer(from, to uuid.UUID, amount float64)
+	Deposit(account uuid.UUID, amount int64) int64
+	Withdraw(account uuid.UUID, amount int64) int64
+	Balance(account uuid.UUID) int64
+	Transfer(from, to uuid.UUID, amount int64)
 	Close() error
 }
